+++ packages/db/migrations/0089_modern_mordo.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "tambo_users" ADD COLUMN "referral_source" text;